project-root/server/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
python-dotenv==1.0.0
anthropic==0.3.0
openai==0.27.0
pydantic==1.10.0
uvicorn[standard]==0.21.0
httpx==0.24.0
typing-extensions==4.5.0 